refactor(file): Move FileMetadata to its own file.
